Vision-RAG/visionrag-backend/pyproject.toml
[project]
name = "visionrag-backend"
version = "0.1.0"
description = "MultiModal RAG"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "torch>=2.0.0",
    "torchvision>=0.17.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.10",
]
